+++ Patches/StartOfRoundPatch.cs
@@ -50,5 +50,57 @@ namespace HeavyItemSCPs.Patches
                 }
             }
         }
+
+        [HarmonyPostfix]
+        [HarmonyPatch(nameof(StartOfRound.OnShipLandedMiscEvents))]
+        public static void OnShipLandedMiscEventsPostfix() // TEST THIS
+        {
+            foreach (var scpItem in SCPItems.SCPItemsList)
+            {
+                scpItem.itemsSpawnedInRound = 0;
+            }
+
+            logger.LogDebug("Checking for Keter SCPs");
+
+            List<string> KeterList = new List<string>();
+
+            foreach (var item in UnityEngine.Object.FindObjectsOfType<GrabbableObject>())
+            {
+                foreach (SCPItem scpItem in SCPItems.SCPItemsList)
+                {
+                    if (scpItem.item == item.itemProperties && scpItem.MaxSpawns != 0)
+                    {
+                        if (scpItem.itemsSpawnedInRound < scpItem.MaxSpawns)
+                        {
+                            scpItem.itemsSpawnedInRound++;
+                        }
+                        else
+                        {
+                            if (!item.hasBeenHeld && !item.isInShipRoom)
+                            {
+                                NetworkObject networkObject = item.gameObject.GetComponent<NetworkObject>();
+
+                                if (networkObject != null && networkObject.IsSpawned && (NetworkManager.Singleton.IsHost || NetworkManager.Singleton.IsServer))
+                                {
+                                    item.gameObject.GetComponent<NetworkObject>().Despawn(true);
+                                }
+                                else
+                                {
+                                    logger.LogDebug("Item prop '" + item.gameObject.name + "' was not spawned, did not have a NetworkObject component or wasnt the server! Skipped despawning and destroyed it instead.");
+                                    UnityEngine.Object.Destroy(item.gameObject);
+                                }
+                            }
+                        }
+
+                        if (scpItem.ObjectClass == ObjectClass.Keter)
+                        {
+                            KeterList.Add(item.itemProperties.itemName);
+                        }
+                    }
+                }
+            }
+
+            if (KeterList.Count > 0) { HUDManager.Instance.DisplayTip("Warning", $"{KeterList.Count} anomalies of class Keter detected:\n{string.Join(", ", KeterList)}", true); }
+        }
     }
 }